organize stuff that was in the categories folder, moved to better place
