--- packages/augur-ui/src/modules/categories/icons.tsx
@@ -1,259 +1,0 @@
-import * as React from "react";
-
-export const Sports = (
-  <svg viewBox="0 0 48 48" fill="none">
-    <path d="M15.9299 3.54797L23.9999 9.59997L32.0699 3.54797" strokeWidth="2" strokeMiterlimit="10"/>
-    <path d="M2.05591 25.354L10.3039 19.554L7.04191 10" strokeWidth="2" strokeMiterlimit="10"/>
-    <path d="M18.506 45.288L15.536 35.65L5.44995 35.8" strokeWidth="2" strokeMiterlimit="10"/>
-    <path d="M42.5499 35.8L32.4639 35.648L29.4939 45.286" strokeWidth="2" strokeMiterlimit="10"/>
-    <path d="M40.958 10L37.696 19.55L45.944 25.35" strokeWidth="2" strokeMiterlimit="10"/>
-    <path d="M27.9999 30L30.9659 22.33L23.9999 16.656L17.0339 22.33L19.9999 30H27.9999Z" strokeWidth="2" strokeMiterlimit="10" strokeLinecap="square"/>
-    <path d="M24 46C36.1503 46 46 36.1503 46 24C46 11.8497 36.1503 2 24 2C11.8497 2 2 11.8497 2 24C2 36.1503 11.8497 46 24 46Z" strokeWidth="2" strokeMiterlimit="10" strokeLinecap="square"/>
-  </svg>
-);
-
-export const Politics = (
-  <svg viewBox="0 0 48 48" fill="none">
-    <path d="M4 24V46" strokeWidth="2" strokeMiterlimit="10"/>
-    <path d="M44 24V46" strokeWidth="2" strokeMiterlimit="10"/>
-    <path d="M34 24V46" strokeWidth="2" strokeMiterlimit="10"/>
-    <path d="M24 24V46" strokeWidth="2" strokeMiterlimit="10"/>
-    <path d="M14 24V46" strokeWidth="2" strokeMiterlimit="10"/>
-    <path d="M2 46H46" strokeWidth="2" strokeMiterlimit="10" strokeLinecap="square"/>
-    <path d="M2 24H46" strokeWidth="2" strokeMiterlimit="10" strokeLinecap="square"/>
-    <path d="M40 24C40 15.2 32.8 8 24 8C15.2 8 8 15.2 8 24" strokeWidth="2" strokeMiterlimit="10" strokeLinecap="square"/>
-    <path d="M24 2V8" strokeWidth="2" strokeMiterlimit="10" strokeLinecap="square"/>
-  </svg>
-);
-
-export const Finance = (
-  <svg viewBox="0 0 48 48" fill="none">
-    <path d="M46.5 1.5H1.5V46.5H46.5V1.5Z" strokeWidth="1.5" strokeMiterlimit="10" strokeLinecap="square"/>
-    <path d="M9 30L19.5 18L30 25.5L39 18" strokeWidth="1.5" strokeMiterlimit="10" strokeLinecap="square"/>
-  </svg>
-);
-
-export const Crypto = (
-  <svg viewBox="0 0 48 48" fill="none">
-    <path d="M7.5 24H9C12.402 24 16.239 25.8705 18 28.5H24C27.0105 28.5 30 31.4895 30 34.5H16.5" strokeWidth="1.5" strokeMiterlimit="10"/>
-    <path d="M29.0415 31.5H39C44.4375 31.5 45 34.5 45 34.5L23.52 43.224C21.8115 43.908 19.8915 43.8285 18.2445 43.005L7.5 37.548" strokeWidth="1.5" strokeMiterlimit="10"/>
-    <path d="M36 24C39.3137 24 42 21.3137 42 18C42 14.6863 39.3137 12 36 12C32.6863 12 30 14.6863 30 18C30 21.3137 32.6863 24 36 24Z" strokeWidth="1.5" strokeMiterlimit="10" strokeLinecap="square"/>
-    <path d="M7.5 21H1.5V40.5H7.5V21Z" strokeWidth="1.5" strokeMiterlimit="10" strokeLinecap="square"/>
-    <path d="M18 15C21.3137 15 24 12.3137 24 9C24 5.68629 21.3137 3 18 3C14.6863 3 12 5.68629 12 9C12 12.3137 14.6863 15 18 15Z" strokeWidth="1.5" strokeMiterlimit="10" strokeLinecap="square"/>
-  </svg>
-);
-
-export const Entertainment = (
-  <svg viewBox="0 0 48 48" fill="none">
-    <path d="M42 15H6L10.5 45H37.5L42 15Z" strokeWidth="1.5" strokeMiterlimit="10" strokeLinecap="square"/>
-    <path d="M39 9.00003C39.0062 7.85473 38.6828 6.73183 38.0683 5.76533C37.4538 4.79883 36.5741 4.02955 35.5344 3.54929C34.4946 3.06902 33.3386 2.89806 32.2044 3.0568C31.0701 3.21554 30.0055 3.69728 29.1375 4.44453C28.6126 3.54806 27.8622 2.80455 26.9609 2.28789C26.0597 1.77123 25.0389 1.49939 24 1.49939C22.9612 1.49939 21.9404 1.77123 21.0391 2.28789C20.1379 2.80455 19.3875 3.54806 18.8625 4.44453C17.9946 3.69728 16.9299 3.21554 15.7957 3.0568C14.6614 2.89806 13.5054 3.06902 12.4657 3.54929C11.4259 4.02955 10.5463 4.79883 9.93179 5.76533C9.31728 6.73183 8.99382 7.85473 9.00003 9.00003" strokeWidth="1.5" strokeMiterlimit="10" strokeLinecap="square"/>
-    <path d="M18 45L16.5 15" strokeWidth="1.5" strokeMiterlimit="10"/>
-    <path d="M30 45L31.5 15" strokeWidth="1.5" strokeMiterlimit="10"/>
-  </svg>
-);
-
-export const World = (
-  <svg viewBox="0 0 48 48" fill="none">
-    <path d="M11.408 5.95801C12.796 6.98401 13.922 8.28601 14.942 9.99801C15.834 11.492 17.816 15.652 16 18C13.946 20.656 8 21.552 8 24C8 25.842 10.608 27.944 12 30C14.094 33.092 13.142 36.088 12 38C11.408 38.992 10.462 39.84 9.414 40.468" strokeWidth="2" strokeMiterlimit="10"/>
-    <path d="M41.336 10.4539C37.018 12.5239 31.084 13.9219 30 13.9999C27.91 14.1499 27.6 12.4319 26 9.99987C24.8 8.17587 22 5.89387 22 3.99987C22 3.25787 22.072 2.65587 22.262 2.06787" strokeWidth="2" strokeMiterlimit="10"/>
-    <path d="M24 46C36.1503 46 46 36.1503 46 24C46 11.8497 36.1503 2 24 2C11.8497 2 2 11.8497 2 24C2 36.1503 11.8497 46 24 46Z" strokeWidth="2" strokeMiterlimit="10" strokeLinecap="square"/>
-    <path d="M38.028 25.8062C38.112 31.9742 30.084 39.6662 26 38.0002C22.42 36.5402 24.946 33.7242 24.028 25.8062C23.646 22.5142 27.162 19.8062 31.028 19.8062C34.894 19.8062 37.984 22.4942 38.028 25.8062Z" strokeWidth="2" strokeMiterlimit="10"/>
-  </svg>
-);
-
-export const USPolitics = (
-  <svg viewBox="0 0 48 48" fill="none">
-    <path d="M8 14C5.79 14 4 12.21 4 10C4 7.79 5.79 6 8 6C10.21 6 12 7.79 12 10C12 12.21 10.21 14 8 14Z" strokeWidth="2" strokeMiterlimit="10" strokeLinecap="square"/>
-    <path d="M12 42H6V34H2V24C2 21.79 3.79 20 6 20H8" strokeWidth="2" strokeMiterlimit="10" strokeLinecap="square"/>
-    <path d="M40 14C42.21 14 44 12.21 44 10C44 7.79 42.21 6 40 6C37.79 6 36 7.79 36 10C36 12.21 37.79 14 40 14Z" strokeWidth="2" strokeMiterlimit="10" strokeLinecap="square"/>
-    <path d="M36 42H42V34H46V24C46 21.79 44.21 20 42 20H40" strokeWidth="2" strokeMiterlimit="10" strokeLinecap="square"/>
-    <path d="M24 14C20.686 14 18 11.314 18 8C18 4.686 20.686 2 24 2C27.314 2 30 4.686 30 8C30 11.314 27.314 14 24 14Z" strokeWidth="2" strokeMiterlimit="10" strokeLinecap="square"/>
-    <path d="M30 46H18V34H14V24C14 21.79 15.79 20 18 20H30C32.21 20 34 21.79 34 24V36H30V46Z" strokeWidth="2" strokeMiterlimit="10" strokeLinecap="square"/>
-  </svg>
-);
-
-export const USElections = (
-  <svg viewBox="0 0 48 48" fill="none">
-    <path d="M18 16L22 20L30 12" strokeWidth="2" strokeMiterlimit="10" strokeLinecap="square"/>
-    <path d="M44 14V40C44 41.0609 43.5786 42.0783 42.8284 42.8284C42.0783 43.5786 41.0609 44 40 44H8C6.93913 44 5.92172 43.5786 5.17157 42.8284C4.42143 42.0783 4 41.0609 4 40V14" strokeWidth="2" strokeMiterlimit="10" strokeLinecap="square"/>
-    <path d="M10 26V4H38V26" strokeWidth="2" strokeMiterlimit="10" strokeLinecap="square"/>
-    <path d="M4 32H44" strokeWidth="2" strokeMiterlimit="10"/>
-  </svg>
-);
-
-export const REP = (
-  <svg viewBox="0 0 48 48" fill="none">
-    <path fillRule="evenodd" clipRule="evenodd" d="M30.3823 26.1203L30.4458 26.4987L30.467 26.6483C30.7353 28.7313 31.5434 29.3401 33.5476 30.2278C33.5518 30.2295 33.5561 30.2287 33.5603 30.2304L40.8309 33.4417L24.4155 42.8142V37.4498C24.4155 35.4815 24.5337 35.1285 26.2801 34.163C26.6233 33.9738 26.9702 33.7758 27.3503 33.5602L27.5875 33.424C28.6746 32.8435 29.5605 31.9846 30.2142 30.9665C30.2379 30.9291 30.4125 30.6656 30.4357 30.628C30.3222 30.5733 30.2021 30.4786 30.0606 30.3551C29.9672 30.2738 29.8283 30.1158 29.7602 29.9859C29.1881 31.0991 28.3194 32.0408 27.2077 32.6535L26.9385 32.8076C26.5636 33.0208 26.1813 33.2296 25.8393 33.4175C23.6977 34.6016 23.5273 35.3272 23.5273 37.4498V42.82L6.96013 33.277L14.0496 30.1766L14.5108 29.9779C14.6912 29.8899 15.2581 29.6583 15.6959 29.6449C16.1427 29.6175 16.8205 29.6462 17.1854 30.1117C17.7036 30.7683 18.1994 31.3957 18.1994 31.3957C18.981 32.4094 20.003 33.2387 21.2241 33.7364C21.2241 33.7364 21.2306 33.7405 21.2348 33.7422C21.5567 33.8739 21.9039 33.9498 22.2477 34.03C22.3184 33.8942 22.4299 33.7362 22.5318 33.608C22.6056 33.5155 22.7161 33.4015 22.8191 33.2947C22.3002 33.1927 21.7029 33.0279 21.2375 32.8046L21.2357 32.8035C20.3084 32.3621 19.5063 31.744 18.89 30.949C18.8808 30.9382 18.8593 30.9125 18.8511 30.9004C18.8511 30.9004 18.8267 30.8666 18.8075 30.8392C18.7453 30.756 18.6825 30.6757 18.6246 30.5897C18.4603 30.3683 18.2113 30.0386 17.8623 29.598C17.6276 29.2997 17.2998 29.1034 16.9509 28.9771V28.9771C16.5744 28.842 16.1401 28.783 15.6655 28.7862C14.9461 28.8098 14.1469 29.1937 14.131 29.2013L13.882 29.3088L13.8627 29.3165L13.2171 29.5948C13.2129 29.5963 13.2095 29.6026 13.2095 29.6026L7.01379 32.3126L23.5123 5.97515V17.1232C23.5123 19.0497 23.3715 19.4704 21.6097 20.4444C21.4747 20.519 21.3314 20.5873 21.1838 20.6569C21.1092 20.693 21.0329 20.7319 20.9565 20.7693C20.8986 20.795 20.8443 20.8255 20.7879 20.853L20.7855 20.8538C18.4739 21.9638 16.8597 24.2154 16.7186 26.8431C16.8706 26.8844 17.1002 26.9496 17.2285 26.9886C17.3224 27.0167 17.4642 27.0871 17.584 27.1516C17.6068 24.7233 19.0724 22.6286 21.1896 21.6174V21.6174C21.1955 21.6151 21.2015 21.6148 21.2064 21.6125C21.3338 21.5467 21.457 21.4879 21.5786 21.43C21.7423 21.3517 21.8999 21.2758 22.0514 21.192C22.3675 21.0177 22.6293 20.8538 22.8676 20.693C22.932 20.6508 22.9865 20.6056 23.0453 20.5635C23.1743 20.4699 23.2959 20.3741 23.4015 20.2798C23.4629 20.2267 23.519 20.1763 23.5593 20.1272C24.2646 19.4112 24.4027 18.6052 24.4027 17.1232V5.96996L40.9525 32.5517L33.6482 29.3256H33.6441C32.0241 28.6085 31.5318 28.0092 31.3145 26.2861C31.3114 26.2626 31.3047 26.204 31.2989 26.1819C30.8815 23.4263 28.8237 21.2153 26.1207 20.4422C25.9361 20.3895 25.7908 20.3462 25.7559 20.3367C25.7002 20.5135 25.6204 20.664 25.5755 20.741C25.4812 20.9018 25.385 21.0138 25.2871 21.1217C25.3285 21.131 25.5862 21.1868 25.6271 21.1967C28.0574 21.7919 29.937 23.7035 30.3823 26.1203ZM42.3218 33.0848L24.3424 4.20776C24.2628 4.07891 24.1184 4 23.9626 4C23.8065 4 23.6623 4.07891 23.5828 4.20711L5.56383 32.9691C5.5028 33.0664 5.48425 33.1847 5.51375 33.2954C5.54125 33.4061 5.61413 33.5018 5.71652 33.5604L23.7354 43.9397C23.805 43.9799 23.8848 44 23.9626 44C24.0406 44 24.1184 43.9799 24.1881 43.9403L42.1675 33.6761C42.2699 33.6173 42.3419 33.5226 42.3723 33.4119C42.4016 33.3003 42.3839 33.183 42.3218 33.0848Z" fill="white"/>
-  </svg>
-);
-
-export const LTC = (
-  <svg viewBox="0 0 48 48" fill="none">
-    <path d="M21.1971 36.0086L21.0302 36.637H21.6803H41.5839L39.7414 43.5H7.84785L11.043 31.4946L11.2899 30.567L10.3883 30.8964L6.52949 32.3059L7.79037 27.5577L12.309 25.9069L12.5536 25.8176L12.6206 25.5659L18.2269 4.5H29.5648L25.3691 20.2984L25.1229 21.2256L26.0239 20.8964L31.277 18.9773L30.0132 23.7265L24.1068 25.8844L23.862 25.9738L23.7951 26.2257L21.1971 36.0086Z" />
-  </svg>
-);
-
-export const ETH = (
-  <svg viewBox="0 0 48 48" fill="none">
-    <path d="M12.0546 24L24 17.5679L35.9454 24L24 30.4321L12.0546 24Z" />
-    <path d="M23.7629 33.4402L24 33.5679L24.2371 33.4402L35.4256 27.4157L24 44.1145L12.5744 27.4157L23.7629 33.4402Z" />
-    <path d="M24 30.4321L11.7047 23.8116L24 3.94993L36.2953 23.8116L24 30.4321Z" />
-    <path d="M24 6V29M24 34V43" />
-  </svg>
-);
-
-export const BTC = (
-  <svg viewBox="0 0 48 48" fill="none">
-    <path d="M31.3571 9.84263V10.2652L31.7738 10.3356C35.5857 10.9802 38.5 14.2968 38.5 18.2857C38.5 20.3531 37.713 22.2342 36.4278 23.666L36.1281 24L36.4278 24.334C37.713 25.7658 38.5 27.6469 38.5 29.7143C38.5 33.7047 35.5857 37.0212 31.774 37.6643L31.3571 37.7347V38.1574V43.5H26.6429V38.2857V37.7857H26.1429H20.4286H19.9286V38.2857V43.5H15.2143V38.2857V37.7857H14.7143H9.5V33.0714H14.7143H15.2143V32.5714V15.4286V14.9286H14.7143H9.5V10.2143H14.7143H15.2143V9.71429V4.5H19.9286V9.71429V10.2143H20.4286H26.1429H26.6429V9.71429V4.5H31.3571V9.84263ZM20.4286 14.9286H19.9286V15.4286V21.1429V21.6429H20.4286H30.4286C32.2805 21.6429 33.7857 20.1376 33.7857 18.2857C33.7857 16.4338 32.2805 14.9286 30.4286 14.9286H20.4286ZM19.9286 32.5714V33.0714H20.4286H30.4286C32.2805 33.0714 33.7857 31.5662 33.7857 29.7143C33.7857 27.8624 32.2805 26.3571 30.4286 26.3571H20.4286H19.9286V26.8571V32.5714Z" />
-  </svg>
-);
-
-export const TV = (
-  <svg viewBox="0 0 48 48" fill="none">
-    <path d="M18 6L24 14" strokeWidth="2" strokeMiterlimit="10"/>
-    <path d="M24 14L34 2" strokeWidth="2" strokeMiterlimit="10"/>
-    <path d="M46 14H2V46H46V14Z" strokeWidth="2" strokeMiterlimit="10" strokeLinecap="square"/>
-    <path d="M30 22H10V38H30V22Z" strokeWidth="2" strokeMiterlimit="10" strokeLinecap="square"/>
-    <path d="M38 32V38" strokeWidth="2" strokeMiterlimit="10" strokeLinecap="square"/>
-    <path d="M38 22V24" strokeWidth="2" strokeMiterlimit="10" strokeLinecap="square"/>
-  </svg>
-);
-
-export const Music = (
-  <svg viewBox="0 0 48 48" fill="none">
-    <path d="M16 14H46" strokeWidth="2" strokeMiterlimit="10" strokeLinecap="square"/>
-    <path d="M46 35V2H16V39" strokeWidth="2" strokeMiterlimit="10"/>
-    <path d="M9 46C12.866 46 16 42.866 16 39C16 35.134 12.866 32 9 32C5.13401 32 2 35.134 2 39C2 42.866 5.13401 46 9 46Z" strokeWidth="2" strokeMiterlimit="10" strokeLinecap="square"/>
-    <path d="M39 42C42.866 42 46 38.866 46 35C46 31.134 42.866 28 39 28C35.134 28 32 31.134 32 35C32 38.866 35.134 42 39 42Z" strokeWidth="2" strokeMiterlimit="10" strokeLinecap="square"/>
-  </svg>
-);
-
-export const Movies = (
-  <svg viewBox="0 0 48 48" fill="none">
-    <path d="M46 42L32 36V28L46 22V42Z" strokeWidth="2" strokeMiterlimit="10" strokeLinecap="square"/>
-    <path d="M32 20H2V44H32V20Z" strokeWidth="2" strokeMiterlimit="10" strokeLinecap="square"/>
-    <path d="M8 14C11.3137 14 14 11.3137 14 8C14 4.68629 11.3137 2 8 2C4.68629 2 2 4.68629 2 8C2 11.3137 4.68629 14 8 14Z" strokeWidth="2" strokeMiterlimit="10" strokeLinecap="square"/>
-    <path d="M24 14C26.2091 14 28 12.2091 28 10C28 7.79086 26.2091 6 24 6C21.7909 6 20 7.79086 20 10C20 12.2091 21.7909 14 24 14Z" strokeWidth="2" strokeMiterlimit="10" strokeLinecap="square"/>
-  </svg>
-);
-
-export const Awards = (
-  <svg viewBox="0 0 48 48" fill="none">
-    <path d="M12 18H10C8.4087 18 6.88258 17.3679 5.75736 16.2426C4.63214 15.1174 4 13.5913 4 12V2H12" strokeWidth="2" strokeMiterlimit="10"/>
-    <path d="M36 18H38C39.5913 18 41.1174 17.3679 42.2426 16.2426C43.3679 15.1174 44 13.5913 44 12V2H36" strokeWidth="2" strokeMiterlimit="10"/>
-    <path d="M24 38V30" strokeWidth="2" strokeMiterlimit="10"/>
-    <path d="M24 30C20.8174 30 17.7652 28.7357 15.5147 26.4853C13.2643 24.2348 12 21.1826 12 18V2H36V18C36 21.1826 34.7357 24.2348 32.4853 26.4853C30.2348 28.7357 27.1826 30 24 30Z" strokeWidth="2" strokeMiterlimit="10" strokeLinecap="square"/>
-    <path d="M36 46H12C12 43.8783 12.8429 41.8434 14.3431 40.3431C15.8434 38.8429 17.8783 38 20 38H28C30.1217 38 32.1566 38.8429 33.6569 40.3431C35.1571 41.8434 36 43.8783 36 46Z" strokeWidth="2" strokeMiterlimit="10" strokeLinecap="square"/>
-  </svg>
-);
-
-export const Indexes = (
-  <svg viewBox="0 0 48 48" fill="none">
-    <path d="M32 32V44" strokeWidth="2" strokeMiterlimit="10" strokeLinecap="square"/>
-    <path d="M32 4V16" strokeWidth="2" strokeMiterlimit="10"/>
-    <path d="M35.99 20.28L46 26" strokeWidth="2" strokeMiterlimit="10"/>
-    <path d="M2 30L12 22L20 28L28.176 21.186" strokeWidth="2" strokeMiterlimit="10"/>
-    <path d="M46 4H2V44H46V4Z" strokeWidth="2" strokeMiterlimit="10" strokeLinecap="square"/>
-    <path d="M32 24C34.2091 24 36 22.2091 36 20C36 17.7909 34.2091 16 32 16C29.7909 16 28 17.7909 28 20C28 22.2091 29.7909 24 32 24Z" strokeWidth="2" strokeMiterlimit="10" strokeLinecap="square"/>
-  </svg>
-);
-
-export const Commodities = (
-  <svg viewBox="0 0 48 48" fill="none">
-    <path d="M24 46V38" strokeWidth="2" strokeMiterlimit="10" strokeLinecap="square"/>
-    <path d="M10 22H12C15.1826 22 18.2348 23.2643 20.4853 25.5147C22.7357 27.7652 24 30.8174 24 34V38H22C18.8174 38 15.7652 36.7357 13.5147 34.4853C11.2643 32.2348 10 29.1826 10 26V22Z" strokeWidth="2" strokeMiterlimit="10" strokeLinecap="square"/>
-    <path d="M26 38H24V34C24 30.8174 25.2643 27.7652 27.5147 25.5147C29.7652 23.2643 32.8174 22 36 22H38V26C38 29.1826 36.7357 32.2348 34.4853 34.4853C32.2348 36.7357 29.1826 38 26 38V38Z" strokeWidth="2" strokeMiterlimit="10" strokeLinecap="square"/>
-    <path d="M20 12C20.089 10.4649 20.4895 8.96394 21.1772 7.58858C21.8648 6.21322 22.8253 4.99224 24 4C25.1747 4.99224 26.1352 6.21322 26.8228 7.58858C27.5105 8.96394 27.911 10.4649 28 12C27.911 13.5351 27.5105 15.0361 26.8228 16.4114C26.1352 17.7868 25.1747 19.0078 24 20C22.8253 19.0078 21.8648 17.7868 21.1772 16.4114C20.4895 15.0361 20.089 13.5351 20 12V12Z" strokeWidth="2" strokeMiterlimit="10" strokeLinecap="square"/>
-  </svg>
-);
-
-export const Stocks = (
-  <svg viewBox="0 0 48 48" fill="none">
-    <path d="M28 28H20V46H28V28Z" strokeWidth="2" strokeMiterlimit="10" strokeLinecap="square"/>
-    <path d="M10 36H2V46H10V36Z" strokeWidth="2" strokeMiterlimit="10" strokeLinecap="square"/>
-    <path d="M46 20H38V46H46V20Z" strokeWidth="2" strokeMiterlimit="10" strokeLinecap="square"/>
-    <path d="M6 20L16 10L24 18L40 2" strokeWidth="2" strokeMiterlimit="10"/>
-    <path d="M30 2H40V12" strokeWidth="2" strokeMiterlimit="10" strokeLinecap="square"/>
-  </svg>
-);
-
-export const HorseRacing = (
-  <svg viewBox="0 0 48 48" fill="none">
-    <path d="M12 44C16 36 26 34 26 26V16C26 18.1218 25.1571 20.1566 23.6569 21.6569C22.1566 23.1572 20.1217 24 18 24L9 30L5 26L16 10V4.00004C20.726 3.99196 25.3769 5.18228 29.5179 7.45975C33.659 9.73722 37.1549 13.0275 39.6791 17.0229C42.2032 21.0184 43.6731 25.5886 43.9513 30.3065C44.2295 35.0243 43.307 39.7356 41.27 44" strokeWidth="2" strokeMiterlimit="10" strokeLinecap="square"/>
-  </svg>
-);
-
-export const Hockey = (
-  <svg viewBox="0 0 48 48" fill="none">
-    <path d="M40.9475 33.3777L26.6706 36.7078L9.36823 5.94742C8.31361 4.07251 5.93868 3.40766 4.0639 4.4625L4.0639 4.46251C2.18655 5.51879 1.52318 7.8984 2.58331 9.77357L21.359 42.9841C21.7138 43.6117 22.379 43.9998 23.1 43.9998H44C45.1045 43.9998 46 43.1043 46 41.9998V37.2367C46 34.5992 43.4921 32.6837 40.9475 33.3777Z" strokeWidth="2" strokeMiterlimit="10" strokeLinecap="square"/>
-    <path d="M28 17V20.625C28 22.2819 31.134 23.625 35 23.625C38.866 23.625 42 22.2819 42 20.625V17" strokeWidth="2" strokeMiterlimit="10"/>
-    <path d="M35 20C38.866 20 42 18.6569 42 17C42 15.3431 38.866 14 35 14C31.134 14 28 15.3431 28 17C28 18.6569 31.134 20 35 20Z" strokeWidth="2" strokeMiterlimit="10" strokeLinecap="square"/>
-  </svg>
-);
-
-export const Tennis = (
-  <svg viewBox="0 0 48 48" fill="none">
-    <path d="M8.59998 39.6978C10.6918 37.6538 12.3533 35.2119 13.4868 32.5159C14.6203 29.8199 15.2028 26.9243 15.2 23.9998C15.2028 21.0752 14.6203 18.1796 13.4868 15.4836C12.3533 12.7876 10.6918 10.3457 8.59998 8.30176" strokeWidth="2" strokeMiterlimit="10"/>
-    <path d="M39.4 39.6979C37.3098 37.6522 35.6491 35.2097 34.5154 32.5137C33.3816 29.8178 32.7976 26.9225 32.7976 23.9979C32.7976 21.0732 33.3816 18.178 34.5154 15.482C35.6491 12.786 37.3098 10.3435 39.4 8.29785" strokeWidth="2" strokeMiterlimit="10"/>
-    <path d="M24 46C36.1503 46 46 36.1503 46 24C46 11.8497 36.1503 2 24 2C11.8497 2 2 11.8497 2 24C2 36.1503 11.8497 46 24 46Z" strokeWidth="2" strokeMiterlimit="10" strokeLinecap="square"/>
-  </svg>
-);
-
-export const Basketball = (
-  <svg viewBox="0 0 48 48" fill="none">
-    <path d="M40 9C36.2 13 34 18.2 34 24C34 29.8 36.2 35.2 40 39" strokeWidth="2" strokeMiterlimit="10"/>
-    <path d="M8 9C11.8 13 14 18.2 14 24C14 29.8 11.8 35.2 8 39" strokeWidth="2" strokeMiterlimit="10"/>
-    <path d="M24 2V46" strokeWidth="2" strokeMiterlimit="10"/>
-    <path d="M46 24H2" strokeWidth="2" strokeMiterlimit="10"/>
-    <path d="M24 46C36.1503 46 46 36.1503 46 24C46 11.8497 36.1503 2 24 2C11.8497 2 2 11.8497 2 24C2 36.1503 11.8497 46 24 46Z" strokeWidth="2" strokeMiterlimit="10" strokeLinecap="square"/>
-  </svg>
-);
-
-export const Golf = (
-  <svg viewBox="0 0 48 48" fill="none">
-    <path d="M40 16.0002H26V12.3538C26 10.9011 26.7876 9.56264 28.0574 8.85716L38.0574 3.30161C40.7236 1.82042 44 3.7483 44 6.79823V12.0002C44 14.2093 42.2091 16.0002 40 16.0002Z" strokeWidth="2" strokeMiterlimit="10" strokeLinecap="square"/>
-    <path d="M26 16V28" strokeWidth="2" strokeMiterlimit="10" strokeLinecap="square"/>
-    <path d="M6 20.0003H18V15.3191C18 14.5191 17.5232 13.796 16.7878 13.4808L6.78784 9.19511C5.4681 8.62951 4 9.59757 4 11.0334V18.0003C4 19.1049 4.89543 20.0003 6 20.0003Z" strokeWidth="2" strokeMiterlimit="10" strokeLinecap="square"/>
-    <path d="M18 20V28" strokeWidth="2" strokeMiterlimit="10" strokeLinecap="square"/>
-    <path d="M36 28H8V36H36V28Z" strokeWidth="2" strokeMiterlimit="10" strokeLinecap="square"/>
-    <path d="M32 36V46" strokeWidth="2" strokeMiterlimit="10" strokeLinecap="square"/>
-    <path d="M12 46V36" strokeWidth="2" strokeMiterlimit="10" strokeLinecap="square"/>
-  </svg>
-);
-
-export const Baseball = (
-  <svg viewBox="0 0 48 48" fill="none">
-    <path d="M20 24H12" strokeWidth="2" strokeMiterlimit="10" strokeLinecap="square"/>
-    <path d="M17.564 13.012L10.314 16.392" strokeWidth="2" strokeMiterlimit="10" strokeLinecap="square"/>
-    <path d="M10.314 31.6079L17.564 34.9879" strokeWidth="2" strokeMiterlimit="10" strokeLinecap="square"/>
-    <path d="M9 40.062C11.2088 38.0101 12.9705 35.5247 14.1749 32.7609C15.3794 29.9972 16.0007 27.0148 16 24C16.0007 20.9852 15.3794 18.0027 14.1749 15.239C12.9705 12.4753 11.2088 9.98983 9 7.93799" strokeWidth="2" strokeMiterlimit="10"/>
-    <path d="M36 24H28" strokeWidth="2" strokeMiterlimit="10" strokeLinecap="square"/>
-    <path d="M37.686 16.392L30.436 13.012" strokeWidth="2" strokeMiterlimit="10" strokeLinecap="square"/>
-    <path d="M30.436 34.9879L37.686 31.6079" strokeWidth="2" strokeMiterlimit="10" strokeLinecap="square"/>
-    <path d="M39 40.062C36.7918 38.0097 35.0305 35.5241 33.8261 32.7605C32.6217 29.9969 32.0001 27.0146 32.0001 24C32.0001 20.9853 32.6217 18.0031 33.8261 15.2395C35.0305 12.4759 36.7918 9.99029 39 7.93799" strokeWidth="2" strokeMiterlimit="10"/>
-    <path d="M24 46C36.1503 46 46 36.1503 46 24C46 11.8497 36.1503 2 24 2C11.8497 2 2 11.8497 2 24C2 36.1503 11.8497 46 24 46Z" strokeWidth="2" strokeMiterlimit="10" strokeLinecap="square"/>
-  </svg>
-);
-
-export const AmericanFootball = (
-  <svg viewBox="0 0 48 48" fill="none">
-    <path d="M5.04395 26.948L21.0579 42.942" strokeWidth="2" strokeMiterlimit="10"/>
-    <path d="M27.0179 5.04395L42.9539 20.9539" strokeWidth="2" strokeMiterlimit="10"/>
-    <path d="M6.18197 41.8C2.12397 37.746 3.35197 22.7 13.044 13.022C22.644 3.42204 37.798 2.11604 41.858 6.17004C45.918 10.224 44.516 25.44 34.996 34.948C25.4 44.53 10.356 45.958 6.18197 41.8Z" strokeWidth="2" strokeMiterlimit="10" strokeLinecap="square"/>
-    <path d="M17 23L25 31" strokeWidth="2" strokeMiterlimit="10" strokeLinecap="square"/>
-    <path d="M23 17L31 25" strokeWidth="2" strokeMiterlimit="10" strokeLinecap="square"/>
-    <path d="M18 30L30 18" strokeWidth="2" strokeMiterlimit="10" strokeLinecap="square"/>
-  </svg>
-);
-
-export const Soccer = (
-  <svg viewBox="0 0 48 48" fill="none">
-    <path d="M15.93 3.5481L24 9.6001L32.07 3.5481" strokeWidth="2" strokeMiterlimit="10"/>
-    <path d="M2.05603 25.354L10.304 19.554L7.04203 10" strokeWidth="2" strokeMiterlimit="10"/>
-    <path d="M18.506 45.2879L15.536 35.6499L5.45001 35.7999" strokeWidth="2" strokeMiterlimit="10"/>
-    <path d="M42.55 35.7999L32.464 35.6479L29.494 45.2859" strokeWidth="2" strokeMiterlimit="10"/>
-    <path d="M40.958 10L37.696 19.55L45.944 25.35" strokeWidth="2" strokeMiterlimit="10"/>
-    <path d="M28 30L30.966 22.33L24 16.656L17.034 22.33L20 30H28Z" strokeWidth="2" strokeMiterlimit="10" strokeLinecap="square"/>
-    <path d="M24 46C36.1503 46 46 36.1503 46 24C46 11.8497 36.1503 2 24 2C11.8497 2 2 11.8497 2 24C2 36.1503 11.8497 46 24 46Z" strokeWidth="2" strokeMiterlimit="10" strokeLinecap="square"/>
-  </svg>
-);
