+++ packages/augur-ui/src/modules/create-market/get-template.ts
