+++ packages/augur-ui/src/modules/create-market/set-categories.ts
@@ -1,13 +1,7 @@
 import { CUSTOM } from 'modules/common/constants';
+import { SortedGroup } from 'modules/types';
 
-export interface SortedGroup {
-  value: string;
-  label: string;
-  subGroup?: Array<SortedGroup>;
-  autoCompleteList?: Array<SortedGroup>;
-}
-
-export const setCategories: Array<SortedGroup> = [
+export const setCategories: SortedGroup[] = [
   {
     value: 'Sports',
     label: 'Sports',
